Bugfix: init TAPi18n.subscribe's ready state according to the underlying sub ready state
Before this fix, if the underlying subscription that got created during
the call to TAPi18n.subscribe was already ready, the ready method of
the object returned by TAPi18n.subscribe would state that the subscription
is not ready when checked immediately after its creation. since we updated
the ready state async only after the onReady callback of the underlying
subscription was called.

Thanks to @rijk for reporting this bug!