Add a mechanizem to disable the i18n-db transformation on specific routes
The transformation replaces the document fields with those that exist in
the
i18n subdocument for the current user lang.

There are situations where we don't want the transformation on certain
routes
in which we use 3rd party packages that use the collection but are not
under
our control (translation editor route for example).

Note: it is possible to disable translation for a specific find/findOne operation
by passing `transform: null` as an option.